+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * More code cleanup
     * Fixed `FlexUser` caching
     * Added back missing `page.types` field in system content configuration [admin#1612](https://github.com/getgrav/grav-plugin-admin/issues/1612)
+    * Console commands: add method for invalidating cache
 1. [](#bugfix)
     * Fixed validation for select field type with selectize
     * Fixed validation for boolean toggles